+++ test/samples/ignore-global/main.js
@@ -1,2 +1,3 @@
-export var immediate = typeof global.setImmediate === 'function' ?
-   global.setImmediate : global.setTimeout;
+export { immediate as immediate1 } from './firstpass';
+export { immediate as immediate2 } from './identifier';
+export { immediate as immediate3 } from './this';