+++ test/samples/other-transforms/foo.js
@@ -1,0 +1,3 @@
+var bar = require( './bar.js' );
+
+module.exports = bar + 1;