+++ test/samples/react-apollo/commonjs-bar.js
@@ -1,0 +1,6 @@
+function Bar () {
+	this.x = 42;
+}
+
+exports.__esModule = true;
+exports.default = Bar;